create medieval annual range module
